refactor(DB Connections): Build Snowflake Dynamic Form (#16875)
* moved all non-BasicParameters into own field

* refactored DB Connection Form made ValidatedInput
